Create Pi using Ramanujan's formula!
The function calculatePi() in the code I provided approximates the value of Pi using Ramanujan's formula, by doing these steps:
The function takes an input parameter maxIterations, which represents the number of iterations to be performed in the Ramanujan series.

Inside the function, a variable sum is initialized to 0.0. This variable will hold the cumulative sum of the terms in the series.

A loop is executed from n = 0 to n = maxIterations. Each iteration of this loop calculates the nth term of the Ramanujan series using the helper function calculateRamanujanTerm(n).

The calculated term is added to the sum variable.

Once all the iterations are complete, the function returns the reciprocal of the sum (1.0 / sum), which provides an approximation of Pi based on Ramanujan's formula.
By adjusting the maxIterations parameter, you can control the accuracy of the approximation. A higher value of maxIterations will yield a more precise estimation of Pi, at the cost of increased computation time.